Déplace le code ré-utilisé dans un composant dédié
